fix: ensure type consistency in heating/lighting data
Adjusts factories to enforce type consistency for `pi_heating_demand` and `power` attributes. Updates related tests to reflect these changes.

+++ database/factories/Devices/HeatingFactory.php
@@ -28,7 +28,7 @@ class HeatingFactory extends Factory
             $this->faker->randomFloat(2, 10, 100);
         $piHeatingDemand = $localTemperature >= $occupiedHeatingSetpoint
             ? 0
-            : min(100, round(($occupiedHeatingSetpoint - $localTemperature) * 100));
+            : (int) min(100, round(($occupiedHeatingSetpoint - $localTemperature) * 100));
 
         return [
             'ieee_address' => $this->ieeeAddressFaker(),
